add readme + clean up
